+++ crates/core/src/host/wasm_common/module_host_actor.rs
@@ -10,8 +10,8 @@ use std::time::Duration;
 
 use super::instrumentation::CallTimes;
 use crate::database_logger::{self, SystemLogger};
-use crate::db::datastore::locking_tx_datastore::MutTxId;
-use crate::db::datastore::traits::{IsolationLevel, Program};
+use spacetimedb_datastore::locking_tx_datastore::MutTxId;
+use spacetimedb_datastore::traits::{IsolationLevel, Program};
 use crate::db::db_metrics::DB_METRICS;
 use crate::energy::{EnergyMonitor, EnergyQuanta, ReducerBudget, ReducerFingerprint};
 use crate::execution_context::{self, ReducerContext, Workload};
