+++ crates/core/src/lib.rs
@@ -15,7 +15,6 @@ pub mod client;
 pub mod config;
 pub mod database_logger;
 pub mod estimation;
-pub mod execution_context;
 pub mod host;
 pub mod module_host_context;
 pub mod replica_context;
