+++ crates/datastore/.gitignore
@@ -1,0 +1,1 @@
+flamegraphs
