+++ crates/datastore/README.md
@@ -1,0 +1,3 @@
+> ⚠️ **Unstable Crate** ⚠️
+>
+> The interface of this crate is **not** stable and may change without notice.
