+++ crates/datastore/src/locking_tx_datastore/delete_table.rs
